refactor(graphics): one module per generator, with a contract that enforces itself
Seventeen generators in a 1,717-line file, with their labels in a `META` table
in a different file. Adding one meant editing three places and forgetting the
third was silent — a generator with no META entry rendered under its own id.
That does not survive the next fifteen.

## The module shape

Each generator is now `generators/<name>.ts`, default-exporting
`defineGenerator({ name, label, description, group, defaults, sample, project })`.
Metadata lives beside the code, so a module is one complete generator and there
is no third place to forget. `defineGenerator` is an identity function whose only
job is keeping `S` inferred, so `sample` and `project` are checked against each
other without the structure type ever being written down.

Adding one is a file plus two lines in `index.ts`. `registry.ts` derives
everything else and no longer holds any data of its own.

## The basis, split by responsibility

`shared.ts` was a grab-bag. It is now six modules behind one barrel —
`math`, `motion` (the loop-closure primitives), `noise`, `disorder`, `svg`,
`tiling` — so a generator still has one import line, but a primitive has an
obvious home. Ridge-specific noise constants moved into `ridgeline.ts`, where
they were only ever used.

## Tests, in layers

Five suites, 945 assertions:

- `graphics-shared` — each basis module alone. A bug in `wobble` used to surface
  as seventeen mysterious golden diffs; now it is one red test with a name.
- `graphics-modules` — the contract. Filename equals `name`, ids are kebab-case,
  descriptions fit a caption, `defaults` contain no key that is not a real
  param (a typo there is otherwise silent), every group is non-empty.
- `graphics-svg-regression` — **new**, and the one that was missing. Hashes catch
  everything and explain nothing: a fill change and a rewrite produce the same
  two hex strings. This pins element counts, extents and palette size as
  readable JSON, so a diff says `circle: 240 -> 260`. It also carries the
  validity assertions every future generator inherits — well-formed root,
  balanced groups, no NaN, coordinates near the frame, under 400 KB as a data
  URI at full density.
- `graphics-integration` — **new**. The seam the others miss: module → registry
  → params → SVG → data URI, including that an unknown name degrades rather than
  throwing (talk frontmatter carries these as free text), that an explicit
  `undefined` falls back instead of clobbering a default, and that the
  convenience `render()` agrees with `sample()` + `project()`.

Two findings from writing them. `r2(1.005)` rounds *down*, because `1.005 * 100`
is `100.49999…` in binary — not a bug, now pinned so nobody "fixes" it into
something that moves goldens. And `encodeURIComponent` leaves parentheses alone
while every generator emits `rgba(...)`, so an unquoted `url(data:…)` would
terminate at the first colour's closing paren; both call sites already quote,
and the test now asserts the quote-safe property rather than a paren-free one
that was never true.

## Verified as the move it is

Not one golden hash changed. `generators/README.md` documents the whole
procedure and the nine properties a new generator is signed up to.
