+++ components/graphics/index.ts
@@ -1,7 +1,7 @@
 export { default as AnimatedBackground } from './AnimatedBackground';
 export { default as GeneratedBackground } from './GeneratedBackground';
 export type { GeneratorName } from './generators';
-export { SAMPLED_GENERATORS } from './generators';
+export { GENERATOR_MODULES, SAMPLED_GENERATORS } from './generators';
 export {
   ACCENT_SWATCHES,
   BRUTALIST_ACCENTS,
@@ -12,6 +12,7 @@ export {
 } from './palette';
 export {
   GENERATOR_LIST,
+  generatorsInGroup,
   getGenerator,
   graphicDataUri,
   renderGraphic,
@@ -19,8 +20,10 @@ export {
 } from './registry';
 export type {
   Generator,
+  GeneratorGroup,
+  GeneratorModule,
   GraphicParams,
   RenderFn,
   SampledGenerator,
 } from './types';
-export { BASE_PARAMS } from './types';
+export { BASE_PARAMS, defineGenerator, GENERATOR_GROUPS } from './types';
